refactor(tcp-plug-socket)!: move Command into module
